requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.10
django-debug-toolbar==1.9.1
django-nose==1.4.5
djangorestframework==3.7.3
Faker==0.8.12
idna==2.6
itypes==1.1.0
Jinja2==2.10
jpush==3.2.9
MarkupSafe==1.0
nose==1.3.7
python-dateutil==2.7.2
pytz==2018.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.22
